requirements-all.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

fastapi==0.115.8
uvicorn==0.30.6
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
jinja2==3.1.5
SQLAlchemy[asyncio]==2.0.36
psycopg[binary]==3.2.5
pynvml==11.5.0
websockets==12.0

setuptools>=70.0.0
wheel>=0.43.0

torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

openai-whisper==20240930
